Move aside.ts into utils as itemListDrag.ts and refactored ]
